Move banner to proposals component
