+++ src/components/Proposals/Proposals.js
@@ -1,7 +1,8 @@
-import React from 'react'
+import React, { useCallback, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import { CardLayout, Header, GU } from '@aragon/ui'
 import Layout from '../Layout'
+import ProposalBanner from './ProposalBanner'
 import ProposalCard from './ProposalCard'
 import { useVotes } from '../../hooks/disputable-voting-logic'
 
@@ -13,8 +14,18 @@ const Proposals = React.memo(function Proposals() {
     history.push(`/proposals/${proposalId}`)
   }
 
+  const [bannerClosed, setBannerClosed] = useState(
+    localStorage.getItem('bannerClosed') === 'true'
+  )
+
+  const handleCloseBanner = useCallback(() => {
+    localStorage.setItem('bannerClosed', 'true')
+    setBannerClosed(true)
+  }, [])
+
   return (
     <section>
+      {!bannerClosed && <ProposalBanner onCloseBanner={handleCloseBanner} />}
       <Layout>
         <Header primary="Proposals" />
         {votes ? (
